move interchange::Json into cjson module
